DBZ-20 Extracting RelationalChangeRecordEmitter base class
